add tool calling test
